refactor(tui): extract buildLayout into src/tui/layout.ts
Lift the static chrome construction (root, layout, sidebar, transcript,
status, hint, slash/file/question panels, input box) out of runTui into
a pure builder. runTui now destructures the references it needs from a
single buildLayout(renderer) call.

Behavior is unchanged: same renderable options, same parent/child wiring,
same inputField.focus() on mount, same selectable flags on chrome.
